refactor: move MAX_ARTISTS constant to entities and update imports
