--- src/constants/playlist.ts
@@ -1,1 +1,0 @@
-export const MAX_ARTISTS = 5;
