+++ tsconfig.json
@@ -22,7 +22,6 @@
       "@components/*": ["./src/components/*"],
       "@hooks/*": ["./src/hooks/*"],
       "@styles/*": ["./src/styles/*"],
-      "@constants/*": ["./src/constants/*"],
       "@public/*": ["./public/*"]
     }
   },
